+++ .vscode/settings.json
@@ -1,3 +1,15 @@
 {
-  "typescript.tsdk": "./node_modules/typescript/lib/"
+  "typescript.tsdk": "./node_modules/typescript/lib/",
+  "files.exclude": {
+    "**/.git": true,
+    "**/.svn": true,
+    "**/.hg": true,
+    "**/.DS_Store": true
+  },
+  "search.exclude": {
+    "**/dist": true,
+    "**/coverage": true,
+    "**/node_modules": true,
+    "**/bower_components": true
+  }
 }
