+++ src/analyzer.ts
@@ -1,8 +1,19 @@
-import * as dom from 'dts-dom';
 import * as astqts from 'astq';
 const ASTQ: typeof astqts.ASTQ = astqts as any;
 
-import { InstanceOfResolver, IPropTypes, IProp, IASTNode } from './index';
+import { InstanceOfResolver } from './index';
+import { IPropTypes, IProp } from './deprecated';
+
+interface IASTNode {
+  type: string;
+  loc: Object;
+  [name: string]: any;
+  value?: any;
+  key?: any;
+  expression?: any;
+  id?: any;
+  body?: any;
+}
 
 const defaultInstanceOfResolver: InstanceOfResolver = (_name: string): undefined => undefined;
 
@@ -30,7 +41,6 @@ function getOptionalDocumentation(propertyNode: any): string {
 export function getTypeFromPropType(node: IASTNode, instanceOfResolver = defaultInstanceOfResolver): IProp {
   const result: IProp = {
     type: 'any',
-    type2: 'any',
     optional: true
   };
   if (isNode(node)) {
@@ -39,54 +49,41 @@ export function getTypeFromPropType(node: IASTNode, instanceOfResolver = default
     switch (type.name) {
       case 'any':
         result.type = 'any';
-        result.type2 = 'any';
         break;
       case 'array':
         result.type = (type.arrayType || 'any') + '[]';
-        result.type2 = dom.create.array(type.arrayType2 || 'any');
         break;
       case 'bool':
         result.type = 'boolean';
-        result.type2 = 'boolean';
         break;
       case 'func':
         result.type = '(...args: any[]) => any';
-        result.type2 = dom.create.functionType([
-          dom.create.parameter('args', dom.create.array('any'), dom.ParameterFlags.Rest)], 'any');
         break;
       case 'number':
         result.type = 'number';
-        result.type2 = 'number';
         break;
       case 'object':
         result.type = 'Object';
-        result.type2 = dom.create.namedTypeReference('Object');
         break;
       case 'string':
         result.type = 'string';
-        result.type2 = 'string';
         break;
       case 'node':
         result.type = 'React.ReactNode';
-        result.type2 = dom.create.namedTypeReference('React.ReactNode');
         break;
       case 'element':
         result.type = 'React.ReactElement<any>';
-        result.type2 = dom.create.namedTypeReference('React.ReactElement<any>');
         break;
       case 'union':
         result.type = type.types.map((unionType: string) => unionType).join('|');
-        result.type2 = dom.create.union(type.types2);
         break;
       case 'instanceOf':
         if (type.importPath) {
           result.type = 'typeof ' + type.type;
-          result.type2 = dom.create.typeof(type.type2);
           (result as any).importType = type.type;
           (result as any).importPath = type.importPath;
         } else {
           result.type = 'any';
-          result.type2 = 'any';
         }
         break;
     }
@@ -109,7 +106,6 @@ function getReactPropTypeFromExpression(node: any, instanceOfResolver: InstanceO
         return {
           name: 'instanceOf',
           type: node.arguments[0].name,
-          type2: dom.create.namedTypeReference(node.arguments[0].name),
           importPath: instanceOfResolver(node.arguments[0].name)
         };
       case 'arrayOf':
@@ -117,15 +113,13 @@ function getReactPropTypeFromExpression(node: any, instanceOfResolver: InstanceO
         return {
           name: 'array',
           arrayType: arrayType.type,
-          arrayType2: arrayType.type2
         };
       case 'oneOfType':
         const unionTypes = node.arguments[0].elements.map((element: IASTNode) =>
           getTypeFromPropType(element, instanceOfResolver));
         return {
           name: 'union',
-          types: unionTypes.map((type: any) => type.type),
-          types2: unionTypes.map((type: any) => type.type2)
+          types: unionTypes.map((type: any) => type.type)
         };
     }
   }
